+++ packages/api/test/e2e/fees.e2e.ts
@@ -73,37 +73,6 @@ describe('fees in cennznet', () => {
             });
         });
     });
-
-    // describe('feeExchange for CennznetExtrinsic', () => {
-    //     afterEach(() => {
-    //         jest.setTimeout(5000);
-    //     });
-    //
-    //     it('makes a transfer (sign, then send)', async done => {
-    //         const tradeAssetId = 17008;
-    //         const trader = keyring.bob;
-    //
-    //         const tx = api.tx.genericAsset.transfer(16000, trader.address, 10000);
-    //         tx.addFeeExchangeOpt({
-    //             assetId: tradeAssetId,
-    //             maxPayment: '50000000000000000',
-    //         });
-    //         return tx.signAndSend(trader, ({events, status}) => {
-    //             console.log('Transaction status:', status.type);
-    //             if (status.isFinalized) {
-    //                 console.log('Completed at block hash', status.value.toHex());
-    //                 console.log('Events:');
-    //
-    //                 events.forEach(({phase, event: {data, method, section}}) => {
-    //                     console.log('\t', phase.toString(), `: ${section}.${method}`, data.toString());
-    //                 });
-    //
-    //                 done();
-    //             }
-    //         });
-    //         done();
-    //     });
-    // });
 });
 
 describe('fees in cennznet (Rxjs)', () => {