+++ packages/crml-cennzx-spot/src/derives/liquidityBalance.ts
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ApiInterfaceRx} from '@cennznet/api/polkadot.types';
-import {AnyAssetId} from '@cennznet/crml-generic-asset/types';
+import {ApiInterfaceRx} from '@cennznet/api/types';
+import {AnyAddress, AnyAssetId} from '@cennznet/types/types';
 import {drr} from '@plugnet/api-derive/util/drr';
 import {Balance, Hash} from '@plugnet/types/interfaces';
 import BN from 'bn.js';
 import {Observable} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
-import {AnyAddress} from '../types';
 import {getExchangeKey} from '../utils/utils';
 import {coreAssetId, coreAssetIdAt} from './shared';