+++ packages/crml-cennzx-spot/src/derives/totalLiquidity.ts
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ApiInterfaceRx} from '@cennznet/api/polkadot.types';
-import {AnyAssetId} from '@cennznet/crml-generic-asset/types';
+import {ApiInterfaceRx} from '@cennznet/api/types';
+import {AnyAssetId} from '@cennznet/types/types';
 import {Hash} from '@plugnet/types/interfaces';
 import BN from 'bn.js';
 import {Observable} from 'rxjs';